refactor(db): consolidate optimized queries into domain packages
Move OptimizedXxxQueries and CachedOptimizedQueries from queries/optimized.go
into their respective domain packages:
- locks/optimized.go — GetLockStatus, GetChatLocksOptimized, GetLockStatusCached
- user/optimized.go — GetUserBasicInfo, GetUserBasicInfoCached
- chats/optimized.go — GetChatBasicInfo, GetChatBasicInfoCached
- antiflood/optimized.go — GetAntifloodSettings, GetAntifloodSettingsCached
- antiraid/optimized.go — GetAntiRaidSettings, GetAntiRaidSettingsCached
- filters/optimized.go — GetChatFiltersOptimized, GetChatFiltersCached
- channels/optimized.go — GetChannelSettings, GetChannelSettingsCached

Remove singleton GetOptimizedQueries() and CachedOptimizedQueries god-object.
Delete dead OptimizedBlacklistQueries (zero methods, never called).

Update all repository files to use local functions instead of queries.GetOptimizedQueries().
Update modules/filters.go to use db_filters package for optimized queries.

Benefits:
- Locality: all query behavior per domain lives in one package
- No RWMutex singleton — db.DB read fresh each call
- No nil-DB fallback constructors — checked at function level
- Easier to test — each domain tests independently

Deleted: alita/db/queries/optimized.go, alita/db/queries/optimized_test.go,
alita/db/queries/model_sync_test.go, alita/db/queries/ (empty)

Added: locks/optimized_test.go, user/optimized_test.go
